backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.9.0
